ms_sender/pyproject.toml
[tool.poetry]
name = "ms-sender"
version = "0.1.0"
description = ""
authors = ["Ustinya Golovkina <ugolovkina@ptsecurity.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.2"
pydantic-settings = "^2.4.0"
loguru = "^0.7.2"
uvicorn = "^0.30.6"
pika = "^1.3.2"
aio-pika = "^9.4.3"
sqlalchemy = "^2.0.34"
alembic = "^1.13.2"
asyncpg = "^0.29.0"
greenlet = "^3.1.0"
celery = "^5.4.0"
flower = "^2.0.1"
fastapi-pagination = "^0.12.27"
pyproject-flake8 = "^7.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 150
ignore = ["E203", "F821", "W503"]

[tool.isort]
profile="black"
line_length=150

[tool.pycln]
all = true
